refactor(foundation): reorganize IR files
