+++ packages/@code-like-a-carpenter/foundation-intermediate-representation/src/index.ts
@@ -1,1 +1,5 @@
+export * from './field';
 export * from './intermediate-representation';
+export * from './model';
+export * from './table';
+export * from './types';
