+++ packages/@code-like-a-carpenter/foundation-intermediate-representation/src/types.ts
@@ -1,0 +1,6 @@
+export interface LambdaConfig {
+  readonly memorySize: number;
+  readonly timeout: number;
+}
+
+export type ProjectionType = 'all' | 'keys_only';
